+++ src/test/java/com/acltabontabon/openwealth/services/orderplacement/OrderPlacementServiceTest.java
@@ -6,28 +6,18 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
+import static org.mockito.Mockito.lenient;
 import static org.mockito.Mockito.when;
 
 import com.acltabontabon.openwealth.commons.Constants;
 import com.acltabontabon.openwealth.commons.Result;
 import com.acltabontabon.openwealth.exceptions.FailedRequestException;
-import com.acltabontabon.openwealth.models.orderplacement.FinancialInstrumentDetails;
-import com.acltabontabon.openwealth.models.custodyservices.FinancialInstrumentIdentification;
-import com.acltabontabon.openwealth.models.custodyservices.PlaceOfTrade;
-import com.acltabontabon.openwealth.models.orderplacement.Account;
-import com.acltabontabon.openwealth.types.AmountType;
-import com.acltabontabon.openwealth.models.orderplacement.BulkOrderDetails;
 import com.acltabontabon.openwealth.models.orderplacement.Order;
-import com.acltabontabon.openwealth.models.orderplacement.OrderQuantity;
-import com.acltabontabon.openwealth.models.orderplacement.RequestedAllocation;
 import com.acltabontabon.openwealth.models.orderplacement.RequestedOrder;
 import com.acltabontabon.openwealth.properties.OpenWealthApiProperties.OrderPlacement;
 import com.acltabontabon.openwealth.services.TestFixtures;
-import com.acltabontabon.openwealth.types.AccountType;
-import com.acltabontabon.openwealth.types.ExecutionType;
-import com.acltabontabon.openwealth.types.OrderSide;
-import com.acltabontabon.openwealth.types.SecurityIdentifierType;
-import com.acltabontabon.openwealth.types.TimeInForce;
+import java.util.function.Consumer;
+import java.util.function.Function;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.ArgumentCaptor;
@@ -39,8 +29,6 @@ import org.springframework.core.task.TaskExecutor;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.client.RestClient;
-
-import java.util.function.Consumer;
 
 @ExtendWith(MockitoExtension.class)
 class OrderPlacementServiceTest {
@@ -159,6 +147,94 @@ class OrderPlacementServiceTest {
         Result<Order> result = orderPlacementService.orders()
             .withCorrelationId(correlationId)
             .createNew(requestedOrder)
+            .submit();
+
+        assertTrue(result.isSuccess());
+
+        HttpHeaders headers = new HttpHeaders();
+        headersConsumerCaptor.getValue().accept(headers);
+        assertEquals(correlationId, headers.getFirst(Constants.HEADER_CORRELATION_ID));
+    }
+    @Test
+    void shouldCancelOrder() {
+        Order expectedOrder = Order.builder().build();
+        String clientOrderId = "XR002";
+
+        lenient().when(apiProperties.getOrder())
+            .thenReturn(TestFixtures.MOCK_URL);
+        when(restClient.delete())
+            .thenReturn(uriSpec);
+        when(uriSpec.uri(any(Function.class)))
+            .thenReturn(headersSpec);
+        when(headersSpec.headers(any(Consumer.class)))
+            .thenReturn(headersSpec);
+        when(headersSpec.retrieve())
+            .thenReturn(responseSpec);
+        when(responseSpec.body(Order.class))
+            .thenReturn(expectedOrder);
+
+        Result<Order> result = orderPlacementService.orders()
+            .withClientOrderId(clientOrderId)
+            .cancel()
+            .submit();
+
+        assertTrue(result.isSuccess());
+        assertNotNull(result.getData());
+        assertEquals(expectedOrder, result.getData());
+    }
+
+    @Test
+    void shouldHandleErrorWhenCancelingOrder() {
+        var statusCode = HttpStatus.INTERNAL_SERVER_ERROR;
+        String clientOrderId = "XR002";
+
+        lenient().when(apiProperties.getOrder())
+            .thenReturn(TestFixtures.MOCK_URL);
+        when(restClient.delete())
+            .thenReturn(uriSpec);
+        when(uriSpec.uri(any(Function.class)))
+            .thenReturn(headersSpec);
+        when(headersSpec.headers(any(Consumer.class)))
+            .thenReturn(headersSpec);
+        when(headersSpec.retrieve())
+            .thenReturn(responseSpec);
+        when(responseSpec.body(Order.class))
+            .thenThrow(new FailedRequestException("Failed to delete order", statusCode));
+
+        Result<Order> result = orderPlacementService.orders()
+            .withCorrelationId("test-correlation-id")
+            .withClientOrderId(clientOrderId)
+            .cancel()
+            .submit();
+
+        assertNotNull(result);
+        assertFalse(result.isSuccess());
+        assertEquals("Failed to delete order placement", result.getMessage());
+    }
+
+    @Test
+    void shouldSetCorrelationIdHeaderWhenCancelingOrder() {
+        Order expectedOrder = Order.builder().build();
+        String clientOrderId = "XR002";
+        String correlationId = "test-correlation-id";
+
+        lenient().when(apiProperties.getOrder())
+            .thenReturn(TestFixtures.MOCK_URL);
+        when(restClient.delete())
+            .thenReturn(uriSpec);
+        when(uriSpec.uri(any(Function.class)))
+            .thenReturn(headersSpec);
+        when(headersSpec.headers(headersConsumerCaptor.capture()))
+            .thenReturn(headersSpec);
+        when(headersSpec.retrieve())
+            .thenReturn(responseSpec);
+        when(responseSpec.body(Order.class))
+            .thenReturn(expectedOrder);
+
+        Result<Order> result = orderPlacementService.orders()
+            .withCorrelationId(correlationId)
+            .withClientOrderId(clientOrderId)
+            .cancel()
             .submit();
 
         assertTrue(result.isSuccess());